:recycle: refactor(admin): change name announcement -> notice
